Fix problem with CodeFactor (more class in one file)
